+++ Src/AppLib/VpnHood.AppLib.WebServer/WebServerOptions.cs
@@ -4,5 +4,4 @@ public class WebServerOptions
 {
     public int? DefaultPort { get; init; }
     public Uri? Url { get; init; }
-    public bool ListenOnAllIps { get; init; }
 }