+++ Src/Apps/Connect.Android.Web/MainActivity.cs
@@ -21,8 +21,7 @@ public class MainActivity : AndroidAppMainActivity
     protected override AndroidAppMainActivityHandler CreateMainActivityHandler()
     {
         return new AndroidAppWebViewMainActivityHandler(this, new AndroidMainActivityWebViewOptions {
-            SpaDefaultPort = AppConfigs.Instance.SpaDefaultPort,
-            SpaListenToAllIps = AppConfigs.Instance.SpaListenToAllIps
+            SpaDefaultPort = AppConfigs.Instance.SpaDefaultPort
         });
     }
 }