+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Upcoming
 * Fix: Linux: Client stops after 60 minutes
 * Fix: Linux: Prevent multiple instances
+* Feature: Add remote-access to VpnHood Client by debug commands
 
 # v7.3.752
 * Fix: Unable to close the Quick Launch pop-up on Android